requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
configparser==5.0.2
idna==3.3
influxdb==5.3.1
influxdb-client==1.23.0
msgpack==1.0.3
pkg-resources==0.0.0
pymodbus==2.5.2
Pypubsub==4.0.3
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
RPi.GPIO==0.7.0
Rx==3.2.0
six==1.16.0
urllib3==1.26.7
